conversational/requirements.txt
gspread
openai
oauth2client
websockets
tiktoken
python-dotenv
pandas